requirements.txt
boto3==1.33.2
requests~=2.31.0
Pillow~=10.1.0
botocore~=1.33.2
django-request~=1.6.3
response~=0.5.0
pip~=22.3.1